refactor(preset-umi): 🎨 调整 worker 代码结构
